Draft: Make first addition operator constexpr.
In C++-11 a constexpr function may not even contain local const's. To avoid turning the implementation into an unreadable mess,
I did not go for a literal "translation". Rather:
- moved (an equivalent of) uFullRange() into UFix as maxRange()
- moved (an equivalent of) neededNIExtra() into UFix as a typedef NIadjusted_t (returning "this type, but with NI adjusted according to RANGE)

The end result actually looks quite a bit simpler to me. I hope that isn't because I'm overlooking something.

NOTE: The return type of worktype::asRaw() is the same integer type as previously defined explicitly as return_type.

NOTE: This is a draft, and I only converted a single operator, for now. Once finished, it should be possible to remove at least needed(S)NIExtra(), but
probably also uFullRange() and sFullRange().